requirements.txt
# Boon-Tube-Daemon Requirements
# Python dependencies for monitoring TikTok/YouTube and posting to social platforms
#
# Removed unused packages (not imported in codebase):
# - TikTokApi: Replaced with Playwright network interception
# - discord-webhook: Using requests directly for webhooks  
# - matrix-nio: Using requests directly for Matrix API
# - pyyaml: Not used
# - aiohttp: Not needed, using requests
# - sqlalchemy: Not used for state tracking
# - colorlog: Using standard logging

# Core dependencies
requests==2.32.5
python-dateutil==2.9.0.post0

# YouTube API
google-api-python-client==2.187.0
google-auth==2.41.0  # Compatible with google-auth-oauthlib 1.2.3
google-auth-oauthlib==1.2.3
google-auth-httplib2==0.2.1

# TikTok monitoring via Playwright (no TikTokApi package needed)
playwright==1.55.0

# Social Media Platforms
Mastodon.py==2.1.4
atproto==0.0.63  # Bluesky AT Protocol

# Configuration
python-dotenv==1.2.1

# LLM Integration
google-generativeai==0.8.5  # Google Gemini API

# Secret Management (optional)
doppler-sdk==1.3.0  # Doppler SDK for Python
# AWS Secrets Manager: boto3>=1.28.0
# HashiCorp Vault: hvac>=1.1.0

# Development/Testing (for CI/CD)
pytest>=8.0.0
pytest-cov>=4.1.0
ruff>=0.1.0
bandit>=1.7.0
safety>=3.0.0
